project structure improved, constructors created, minor code changes
