+++ bin/.gitignore
@@ -1,0 +1,10 @@
+/Banco.class
+/Cliente.class
+/Conta.class
+/ContaCorrente.class
+/ContaPoupanca.class
+/IConta.class
+/Main.class
+/main/
+/entities/
+/application/
