--- src/Banco.java
@@ -1,24 +1,0 @@
-import java.util.List;
-
-public class Banco {
-
-	private String nome;
-	private List<Conta> contas;
-
-	public String getNome() {
-		return nome;
-	}
-
-	public void setNome(String nome) {
-		this.nome = nome;
-	}
-
-	public List<Conta> getContas() {
-		return contas;
-	}
-
-	public void setContas(List<Conta> contas) {
-		this.contas = contas;
-	}
-
-}
