+++ src/application/Main.java
@@ -1,3 +1,8 @@
+package application;
+import entities.Cliente;
+import entities.Conta;
+import entities.ContaCorrente;
+import entities.ContaPoupanca;
 
 public class Main {
