+++ src/entities/Cliente.java
@@ -1,7 +1,17 @@
+package entities;
 
 public class Cliente {
 
 	private String nome;
+	
+	public Cliente() {
+		
+	}
+	
+	public Cliente(String nome) {
+		super();
+		this.nome = nome;
+	}
 
 	public String getNome() {
 		return nome;
