+++ src/entities/Conta.java
@@ -1,3 +1,4 @@
+package entities;
 
 public abstract class Conta implements IConta {
