+++ src/entities/ContaCorrente.java
@@ -1,3 +1,4 @@
+package entities;
 
 public class ContaCorrente extends Conta {
