+++ src/entities/ContaPoupanca.java
@@ -1,3 +1,4 @@
+package entities;
 
 public class ContaPoupanca extends Conta {
 
@@ -7,7 +8,7 @@ public class ContaPoupanca extends Conta {
 
 	@Override
 	public void imprimirExtrato() {
-		System.out.println("=== Extrato Conta Poupança ===");
+		System.out.println("=== Extrato Conta Poupanï¿½a ===");
 		super.imprimirInfosComuns();
 	}
 }
