+++ src/entities/IConta.java
@@ -1,3 +1,4 @@
+package entities;
 
 public interface IConta {
